// src/RAAP.Web.Client2/ClientApp/app/models/company.model.ts
export class Company {
    companyId: number;
    name: string;
    createdOn: Date;
    updatedOn: Date;
    databaseName: string;
    profileImage: string|any;
    profileImageFileType: string;
    address1: string;
    address2: string;
    address3: string;
    address4: string;
    phone: string;
    email: string;
    organizationNumber: string;
    homepage: string;
}